+++ app/libraries/Theme.php
@@ -85,6 +85,7 @@ final class Theme extends AbstractChildTheme
     private function setUpStyles()
     {
         $this->setups['Styles\Core'] = new Setups\Styles\Core($this);
+        $this->setups['Styles\Gutenberg'] = new Setups\Styles\Gutenberg($this);
     }
 
     private function setUpScripts()